PDF Refactored
