--- packages/client/src/components/ResumeComponents/pdf/basicTemplate/contact/ContItem.tsx
@@ -1,3 +1,0 @@
-const ContItem = () => {};
-
-export default ContItem;
