+++ packages/client/src/components/ResumeComponents/pdf/templates/basicTemplate/components/HrBar.tsx
