+++ packages/client/src/components/ResumeComponents/pdf/templates/basicTemplate/experience/Experience.tsx
@@ -1,29 +1,6 @@
-import { StyleSheet } from "@react-pdf/renderer";
-import Section from "./../Section";
-import { SectionContainerV } from "./../SectionContainer";
+import Section from "./../components/Section";
+import { SectionContainerV } from "./../components/SectionContainer";
 import ExperienceItem from "./ExperienceItem";
-const styles = StyleSheet.create({
-  sectionTitle: {
-    color: "#336699",
-    fontSize: "15",
-    fontWeight: "bold",
-  },
-  sectionChild1: {
-    fontSize: "16px",
-    fontWeight: "bold",
-  },
-  sectionChild2: {
-    fontSize: "14px",
-    fontWeight: "bold",
-  },
-  sectionChild3: {
-    fontSize: "12px",
-    fontWeight: "bold",
-  },
-  innerView: {
-    marginBottom: 10,
-  },
-});
 
 type jobPropsType = {
   id: string;
