+++ packages/client/src/components/ResumeComponents/pdf/templates/basicTemplate/skill/Skills.tsx
@@ -1,5 +1,5 @@
 import { StyleSheet, View } from "@react-pdf/renderer";
-import Section from "../Section";
+import Section from "../components/Section";
 import SkillItem from "./SkillItem";
 
 const styles = StyleSheet.create({
